+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `ToggleTip`: Add checks for page bottom limit.
+
 ## [12.9.0] - 2022-09-16 👋
 
 Features: